feat(jdbc): centralize connection properties from `BigQueryConnection` and `DataSource` (#4116)

+++ google-cloud-bigquery-jdbc/src/test/java/com/google/cloud/bigquery/jdbc/it/ITBigQueryJDBCTest.java
@@ -1508,7 +1508,7 @@ public class ITBigQueryJDBCTest extends ITBase {
     ds.setURL("jdbc:bigquery://https://www.googleapis.com/bigquery/v2:443;");
     ds.setOAuthType(0);
     ds.setOAuthPvtKeyPath(tempFile.toPath().toString());
-    assertEquals(0, ds.getOAuthType());
+    assertEquals(0, ds.getOAuthType().intValue());
     assertEquals(tempFile.toPath().toString(), ds.getOAuthPvtKeyPath());
   }
